Move Storage-based matrix code to a separate file
